refactor: remove any top-level calls to getAuth()
The global Firebase Auth object is transient. We don't want any
code relying on the values returned by the Auth object's transient
properties (e.g. currentUser). So lets remove all calls to it
at the top-level and setup functions, and instead require all code
that will use it to get a reference to the current instance of the
Auth object.